Demo for AnimationComplete Event
Action: click the red ball, trigger the arm to do a revolution animation.
When the revolution finished, the AnimationComplete event triggers the red ball to do a scale Wave animation.

Purpose: demonstrate AnimationComplete callback event. When an animation completes, it sends out an event that can
be caught with a listener, and then used to do anything. In this case it triggers the animation of the red ball (scale wave).