Simple demo using Event Listenter and Emit()
Action: Click the red sphere to make the green octahedron animate (it wobbles its scale).

Purpose: Illustrate how one object can call another to do something. Specifically, a mouse event on a 3d object (a button, in this case the red sphere) triggers an animation on a separate entity (the green Octahedron).